renamed Parameters.hash -> Parameters.map
