+++ test/map_constraints_test.rb
@@ -26,3 +26,22 @@ describe 'map parameter constraints' do
   rejects(123)
   rejects('abc')
 end
+
+describe 'open-ended map parameter constraints' do
+
+  subject do
+    ActionController::Parameters.map
+  end
+
+  def self.permits(value, options = {})
+    options[:as] ||= value
+    options[:as] = options[:as].with_indifferent_access
+
+    super(value, options)
+  end
+
+  permits(:id => 1, :name => 'Mick')
+  permits({:id => 1, :name => 'Mick'})
+  rejects("a string")
+  rejects(123)
+end
